Move memory list
