+++ src/System.Buffers.Primitives/System/Buffers/IMemoryList.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace System.Collections.Sequences
+namespace System.Buffers
 {
     public interface IMemoryList<T>
     {
