+++ tests/System.IO.Pipelines.Performance.Tests/ReadCursorOperationsThroughput.cs
@@ -128,7 +128,7 @@ namespace System.IO.Pipelines.Performance.Tests
 
                     if (length == -1)
                     {
-                        var subBuffer = buffer.Slice(reader.SequencePosition);
+                        var subBuffer = buffer.Slice(reader.Position);
                         var position = subBuffer.PositionOf((byte)'\n');
                         if (position == null)
                         {
